pyproject.toml
[project]
name = "mission-pack-nodered-bundle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.8",
    "loguru>=0.7.3",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "typer>=0.15.2",
]
